pdf_agent/requirements.txt
phidata
python-dotenv
packaging
fastapi
uvicorn
groq
openai
chromadb
pypdf
